[ParkingLot][Bryan] refactor: remove casting of double in getAvailableCapacity

+++ src/main/java/com/oocl/cultivation/ParkingLot.java
@@ -46,7 +46,7 @@ public class ParkingLot {
     }
 
     double getAvailableCapacity() {
-        return (double) getCapacity() - getParkedCarsSize();
+        return getCapacity() - getParkedCarsSize();
     }
 
     double getAvailablePositionRate() {
